refactor: handler domain 으로 이동
