+++ src/main/java/com/hubo/gillajabi/crawl/domain/service/duru/RoadDuruThemeHandler.java
@@ -1,4 +1,4 @@
-package com.hubo.gillajabi.crawl.application.service;
+package com.hubo.gillajabi.crawl.domain.service.duru;
 
 
 import com.hubo.gillajabi.crawl.application.response.RoadCrawlResponse;
